Fix for when running distributed code:
The previous upgrade to Angular 2 Release Candidate 1 made it work
during development. What I hadn't tested was running the distributed
code, this produced an error relating to reflect-metadata. Removing
the reference to zone.js and reflect-metadata in the index.html and
adding imports for both in the boot.ts file seems to have fixed it,
these must come before anything else.